解决像素差为负时超出uint8范围的问题

+++ Recovery.m
@@ -23,7 +23,7 @@ for i = 1 : m
         for p = x+1 : 1 : x+blocksize-2
             for q = y+1 : 2 : y+blocksize-2
                 l = floor((origin(p,q)+origin(p,q+1))/2);
-                h = double(origin(p,q)-origin(p,q+1));
+                h = double(double(origin(p,q))-double(origin(p,q+1)));
                 min = 2*(255-l);
                 b = 2*l+1;
                 if b < min
@@ -39,6 +39,7 @@ for i = 1 : m
         end
     end
 end
+
 
 %calculate every block's data according to the DATA
 board =zeros(1,m*n);
